--- meoku/src/App.jsx
@@ -1,12 +1,0 @@
-import './App.css'
-
-function App() {
-
-    return (
-        <>
-            test comment
-        </>
-    )
-}
-
-export default App
