--- meoku/src/main.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
-
-ReactDOM.createRoot(document.getElementById('root')).render(
-    <React.StrictMode>
-        <App/>
-    </React.StrictMode>,
-)
